feat(home, public, data): add skills section
This commit adds skills section in homepage. Additionally, it adds an illustration and technology logos in the public directory, along with a tech statck file in the data directory needed for the skills section.

The changes include:

- Create `Skills` component file which contains the implementation of the skills section with a title, a description, tech stack, and an image
- Style the component to be responsive
- Add technology logos in the `public/tech` directory to use them in the skills section
- Create techStack file in the `app/data` directory which defines the list of technologies used in the skills section
- Add SVG image for the thought process illustration used in the skills section
- Modify the root page to import and include the `Skills` component in the homepage

The changes implement a responsive and visually appealing `Skills` section to showcase proficiency in various technologies.