feat(home): add about section
This commit adds a responsive and visually appealing `about` section to the Home page. The section includes a main title, three subsections with their respective titles and descriptions, and an image.

The changes include:
- Create `About` component file which contains the implementation of the about section
- Add a new SVG image in the public folder to use it in the about section
- Style the component to be responsive
- Modify global stylesheet to include utility classes for styling the about subsections
- Update the Home page to import and include the `About` component in the homepage

These changes enhance the user experience by providing a brief summary of my personal life and career.