refactor(projects): update `My Portfolio` project data
This commit refactors the `projects` data file to update the data of the `My Portfolio` project with actual information of the project rather than dummy texts.

The changes include:
- Replace the previous dummy (lorem ipsum) summary with the project's actual summary
- Include subsections with points to provide more details on the project, such as `Features`, `Usage`, and `Future Development Plans`

Additionally, it updates the project title from `Portfolio` to `My Portfolio` so that it becomes more clear, resulting in better user experience.