feat(ProjectDetails): generate page's static parameters
This commit updates the dynamic `Project Details` page to generate its static parameters. The changes include:
- Create a `Section` component and move the entire `<section>` element of the page there in order to separate the client logic from the page
- Generate the static parameters of the page using each of the project's id on the `Project Details` page

Note that the entire `<section>` element of the page (which contains the client elements and client functionality) had to be moved into another component (`Section` component), because the static parameters can only be generated inside a React Server Component (RSC).