fix(Cover): Resolve cover carousel flicker issue
This commit fixes the flicker issue of the cover carosel on the `Project Details` page which was caused by the loading time of each mockup image.

The following changes are made to fix the issue:
- Add a new `isFirstCoverLoaded` state hook which tracks the loading status of the first mockup image
- Only start the carousel cycle animation if the first mockup image has completed loading
- Create a container element for each of the mockup and align them horizontally, instead of just creating a single <Image> component (this mainly fixes the issue as all the image the loaded on the server-side)
- Cycle through the mockup images by moving the entire cover carousel horizontally (instead of changing the source of the image with state hook)
- Remove the `activeCoverId` state hook as it is no longer needed

Additionally, the cycle time for the cover image to change has been reduced to 2 seconds to improve the user experience. Also, the shape and padding of the elements have been adjusted to reflect the changes.