feat(projects): add open graph URLs and update info
This commit adds the openGraphImage property to the projects data file. The openGraphImage property holds the URL of the corresponding project's image to be used for Open Graph meta tags. This will enhance how the project appears when shared on social media platforms or other websites.

Additionally, some minor updates are made to the `My Portfolio` project. The changes include:
- Modify `Features` section points to provide clearer descriptions of the project's features
- Remove the `Screenshots or demo videos showcasing the project in action.` point from the `Usage` section which was accidentally added previously
- Update the link to project live demo