fix: resolve recommended screen scaling issue
This commit fixes the recommended, 125% or 150% scaling option found on the windows 10 and 11 screen resolution setting. A recommended and default scaling above 100% is not unusual on some laptops now-a-days (especially newer laptops with their higher pixel density). Therefore, it creates an issue on the app, specially, on the homepage sections as they are designed to be fit within the full viewport height.

The changes to resolve the issue include:
- Add two custom tailwind breakpoints, `dpr-lg` and `dpr-xl` which activate when user device's screen (landscape orientated) has a minimum device-pixel-ratio of `1.2` and `1.45` respectively
- Adjust the font-size, space, padding, width, and height to make sure the contents fit and do not get cut off in their respective container

Additionally, the commit refactors some small changes. The start and end position of the scroll-trigger for the homepage sections are adjusted to ensure that their corresponding animations are triggered properly. Also, the small desktop screen styles are adjusted to create more whitespace between homepage sections, resulting in improved user interface.